refactor: move shader fragments to separate TU to avoid C++ parsing in tests, update shader tests
test_shader_named_outputs used to parse C++ source files to extract the GLSL shaders in order to validate them with glslang, even though the tests already access other engine code directly. It broke when I introduced the shader proprocessor / fragment system. This fixes the issue by moving the shader code into a separate translation unit (EngineGL33_ShaderSources.cpp) and exposing the preprocessed shader source through a public API.

This commit also adjusts and removes some tests.

Tests that assert that certain enum cases have expected values were removed, because we don't need to test the compiler. The cases were assigned explicit numbers instead. The unit test assertions about the numbers of vertex and pixel shaders were moved to static assertions (though IMHO they could be just removed entirely).
